% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_alfak2}
\alias{plot_alfak2}
\title{Plot alfak2 posterior summaries}
\usage{
plot_alfak2(x, layer = c("global", "local"), ...)
}
\arguments{
\item{x}{An \code{alfak2_fit} or \code{alfak2_local_fit}.}

\item{layer}{\code{"global"} or \code{"local"} for full fits.}

\item{...}{Reserved.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot alfak2 posterior summaries
}
