% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize_alfak2}
\alias{summarize_alfak2}
\title{Summarize alfak2 fits}
\usage{
summarize_alfak2(object, layer = c("global", "local"), ...)
}
\arguments{
\item{object}{An \code{alfak2_fit}, \code{alfak2_local_fit}, or graph posterior list.}

\item{layer}{For full fits, \code{"global"} or \code{"local"}.}

\item{...}{Reserved.}
}
\value{
Data frame of posterior node summaries.
}
\description{
Summarize alfak2 fits
}
