% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-package.R
\docType{package}
\name{alfak2-package}
\alias{alfak2}
\alias{alfak2-package}
\title{alfak2: compiled hierarchical karyotype fitness inference}
\description{
\code{alfak2} fits one inference pipeline for sparse two-timepoint karyotype count
data: a local hierarchical Bayesian posterior implemented in TMB followed by
a graph Gaussian posterior with ordered copy-number epistasis penalties solved
in compiled sparse linear algebra.
}
\author{
\strong{Maintainer}: Tao Li \email{Tao.Li2@moffitt.org}

Authors:
\itemize{
  \item Richard Beck \email{richard.beck@moffitt.org}
}

}
\keyword{internal}
